Input filter error
I'm testing this on both a development server and a production server and the error variable is always returning as false even when it's clear that the err variable is passing data. I believe this is fixed if the single quotes are removed from the first parameter (INPUT_GET) within the input filter.